use Guillaume's request_json function instead of requests
